Fix schema for Problems.rating and test

+++ frontend/database/11_rating_contest_problems.sql
@@ -1,10 +1,10 @@
 ALTER TABLE
 	Contests
 ADD COLUMN
-	`rating` float(2,2) NOT NULL DEFAULT '3.0';
+	`rating` float NOT NULL DEFAULT '3.0';
 
 ALTER TABLE
 	Problems
 ADD COLUMN
-	`rating` float(2,2) NOT NULL DEFAULT '3.0';
+	`rating` float NOT NULL DEFAULT '3.0';
